MOD: double -> float

+++ include/camera.h
@@ -4,16 +4,16 @@
 
 typedef struct s_camera
 {
-	double		aspect_ratio;
+	float		aspect_ratio;
 	int			image_width;
 	int			image_height;
 	t_vec3		origin;
 	t_vec3		origin2;
 	t_vec3		vector;
 	t_vec3		vector2;
-	double		fov;
-	double		vp_height;
-	double		vp_width;
+	float		fov;
+	float		vp_height;
+	float		vp_width;
 	t_vec3		vup;
 	t_vec3		w;
 	t_vec3		u;
@@ -21,15 +21,15 @@ typedef struct s_camera
 	t_vec3		init_w;
 	t_vec3		init_u;
 	t_vec3		init_v;
-	double		focal_length;
+	float		focal_length;
 	int			samples_per_pixel;
 	t_vec3		horizontal;
 	t_vec3		vertical;
 	t_vec3		focal;
 	t_vec3		lower_left_corner;
-	double		alpha;
-	double		beta;
-	double		gamma;
+	float		alpha;
+	float		beta;
+	float		gamma;
 }				t_camera;
 
 #endif
