+++ test/two_sphere.c
@@ -8,7 +8,7 @@ static t_color	ray_color(t_ray* r, t_hlist *world)
 {
 	t_hit_record	rec;
 	t_vec3			unit_direction;
-	double			t;
+	float			t;
 	t_color			tmp1;
 	t_color			tmp2;
 
@@ -24,7 +24,7 @@ static t_color	ray_color(t_ray* r, t_hlist *world)
 int	main()
 {
 	t_mlx			mlx;
-	const double	aspect_ratio = 16.0 / 9.0;
+	const float	aspect_ratio = 16.0 / 9.0;
 	const int		image_width = 1600;
 	const int		image_height = (int)(image_width / aspect_ratio);
 	t_hlist			*world;
@@ -42,9 +42,9 @@ int	main()
 	push(&world, list_(hittable1));
 	push(&world, list_(hittable2));
 
-	double			viewport_height = 2.0;
-	double			viewport_width = aspect_ratio * viewport_height;
-	double			focal_length = 1.0;
+	float			viewport_height = 2.0;
+	float			viewport_width = aspect_ratio * viewport_height;
+	float			focal_length = 1.0;
 
 	t_point3			origin = {0, 0, 0};
 	t_vec3			horizontal = {viewport_width, 0, 0};
@@ -67,8 +67,8 @@ int	main()
 		int i = 0;
 		while (i < image_width)
 		{
-			double	u = (double)i / (image_width - 1);
-			double	v = (double)(image_height - 1 - j) / (image_height - 1);
+			float	u = (float)i / (image_width - 1);
+			float	v = (float)(image_height - 1 - j) / (image_height - 1);
 			t_ray		r;
 			r.origin = origin;
 			r.direction = subtract(add(add(lower_left_corner, multiply(horizontal, u)), multiply(vertical, v)), origin);
